Adding functioning RESTapi
