+++ cocktailMolotov/models.py
@@ -6,18 +6,7 @@ from flask import request, jsonify
 import json
 
 
-# Loop through the ingredients
-# change Ingredients -> only rum (ex) or do if ex: rum in ingredients
-cocktails = [
-    {'id': 0,
-    'name': 'mojito',
-    'ingredients': ['50 ml White Rum', '8 Mint leaves', '12 1/2 ml Sugar Syrup', '25 ml Lime Juice', '2 Mint sprigs'],
-    'description': 'Add the white rum to a highball glass. Add 8 – 10 mint leaves and sugar syrup and lime juice.Muddle with bar spoon. Add crushed ice and a splash of soda. Mix drink down with bar spoon. Taste. Top up with more crushed ice. Slap 2 mint sprigs to release essence and put into drink. Add small splash of soda and straw.'},
-    {'id': 1,
-    'name': 'martini',
-    'ingredients': ['50 ml Gin', '10 ml Dry Vermouth', 'Lemon twist'],
-    'description': 'Chill martini glass with soda water and ice. Fill mixing glass to top with ice. Add the dry vermouth to the mixing glass, giving a small stir to coat the ice with the vermouth. Drain out glass, leaving just the coating on the ice. Add the gin to the mixing glass. Stir for 15 seconds, always making sure that the glass is full to the brim with ice. Taste. Fine strain into chilled martini glass. Zest with lemon peel and add twist unto drink.'}
-]
+
 
 
 
@@ -46,9 +35,7 @@ def db_test():
 def test():
     return 'RUN'
 
-@app.route('/api/v1/ressources/cocktails/all', methods=['GET'])
-def api_all():
-    return jsonify(cocktails)
+
